Use regular expression to get distro postfix.

+++ build.sh
@@ -606,11 +606,11 @@ if [ x"QEMU" != x"$PLATFORM" ] && [ -d $binary_dir ]; then
     fi
 
 	if [ x"HiKey" = x"$PLATFORM" ]; then
-		if [ ! -f $PREBUILD/hisi-idt.py ]; then
+		if [ -f $PREBUILD/hisi-idt.py ] && [ ! -f $binary_dir/hisi-idt.py ]; then
 	    	cp $PREBUILD/hisi-idt.py $binary_dir/
 		fi
 
-		if [ ! -f $PREBUILD/nvme.img ]; then
+		if [ -f $PREBUILD/nvme.img ] && [ ! -f $binary_dir/nvme.img ]; then
 	    	cp $PREBUILD/nvme.img $binary_dir/
 		fi
 	fi
@@ -878,48 +878,40 @@ uncompress_distro()
 	    mkdir -p "$distro_dir" 2> /dev/null
 	    
 	    echo "Uncompressing the distribution($1) ..."
-	    if [ x"${image##*.}" = x"bz2" ] ; then
-	    	TEMP=${image%.*}
-	    	if [ x"${TEMP##*.}" = x"tar" ] ; then
-	    		tar jxvf $DISTRO_DIR/$image -C $distro_dir 2> /dev/null 1>&2
-	    		echo "This is a tar.bz2 package"
-	    	else
-	    		bunzip2 $DISTRO_DIR/$image -C $distro_dir 2> /dev/null 1>&2
-	    		echo "This is a bz2 package"
-	    	fi
-	    fi
-	    if [ x"${image##*.}" = x"gz" ] ; then
-	    	TEMP=${image%.*}
-	    	if [ x"${TEMP##*.}" = x"tar" ] ; then
-	    		sudo tar zxvf $DISTRO_DIR/$image -C $distro_dir 2> /dev/null 1>&2
-	    		echo "This is a tar.gz package"
-	    	else
-	    		gunzip $DISTRO_DIR/$image -C $distro_dir 2> /dev/null 1>&2
-	    		echo "This is a gz package"
-	    	fi
-	    fi
-	    if [ x"${image##*.}" = x"tar" ] ; then 
-	    	tar xvf $DISTRO_DIR/$image -C $distro_dir 2> /dev/null 1>&2
-	    	echo "This is a tar package"
-	    fi
-	    if [ x"${image##*.}" = x"xz" ] ; then 
-	    #	echo "This is a xz package"
-	    	TEMP=${image%.*}
-	    	if [ x"${TEMP##*.}" = x"tar" ] ; then
-	    		xz -d $DISTRO_DIR/$image 2> /dev/null 1>&2
-	    		tar xvf $DISTRO_DIR/$TEMP -C $distro_dir 2> /dev/null 1>&2
-	    	fi
-	    fi
-	    if [ x"${image##*.}" = x"tbz" ] ; then
-	    	tar jxvf $DISTRO_DIR/$image -C $distro_dir 2> /dev/null 1>&2
-	    fi
-	    if [ x"${image}" = x"" ] ; then
-	    	echo "Can not find the suitable root filesystem!"
-	        exit 1
-	    fi
+	    image_postfix=`echo "$image" | grep -Po "((\.tar)*\.(tar|bz2|gz|xz)$)" 2>/dev/null`
+	    case "$image_postfix" in
+		.tar.bz2 )
+			sudo tar zxvf $DISTRO_DIR/$image -C $distro_dir 2> /dev/null 1>&2
+			echo "This is a tar.bz2 package"
+			;;
+		.bz2 )
+			bunzip2 $DISTRO_DIR/$image -C $distro_dir 2> /dev/null 1>&2
+			echo "This is a tar.bz2 package"
+			;;
+		.tar.gz )
+			sudo tar zxvf $DISTRO_DIR/$image -C $distro_dir 2> /dev/null 1>&2
+			echo "This is a tar.gz package"
+			;;
+		.gz )
+			gunzip $DISTRO_DIR/$image -C $distro_dir 2> /dev/null 1>&2
+			echo "This is a tar.gz package"
+			;;
+		.tar.xz | .xz)
+			tar xvf $DISTRO_DIR/$TEMP -C $distro_dir 2> /dev/null 1>&2
+			echo "This is a xz package"
+			;;
+		.tbz )
+			tar jxvf $DISTRO_DIR/$image -C $distro_dir 2> /dev/null 1>&2
+			echo "This is a tbz package"
+			;;
+		* )
+			echo "Can not find the suitable root filesystem!"
+			exit 1
+			;;
+	    esac
 
-		echo "Remove old module files in rootfs..."
-		sudo rm -rf $distro_dir/lib/modules/*
+	    echo "Remove old module files in rootfs..."
+	    sudo rm -rf $distro_dir/lib/modules/*
 
 	fi
 }
@@ -1119,8 +1111,37 @@ install_pkgs()
 ######################### create distribution           ###########################
 ###################################################################################
 distro_postfix=".tar.gz"
+
+rc_local_template="#!/bin/sh -e
+#
+# rc.local
+#
+# This script is executed at the end of each multiuser runlevel.
+# Make sure that the script will \"exit 0\" on success or any other
+# value on error.
+#
+# In order to enable or disable this script just change the execution
+# bits.
+#
+# By default this script does nothing.
+
+exit 0"
+
 create_distro()
 {
+	local rc_local_file=
+	case "$1" in
+		OpenSuse )
+			rc_local_file=rc.d/boot.local
+			;;
+		Fedora )
+			rc_local_file=rc.d/rc.local
+			;;
+		* )
+			rc_local_file=rc.local
+			;;
+	esac
+
 	distro_dir=${PRJROOT}/$build_dir/$DISTRO_DIR/$1
 	image="$1_$TARGETARCH$distro_postfix"
 	if [ x"" != x"$1" ] && [ x"" != x"$image" ] && [ ! -f "$build_dir/$DISTRO_DIR/$image" ]; then
@@ -1132,9 +1153,20 @@ create_distro()
 		sed -i "s/lastupdate=.*/lastupdate=\"$lastupdate\"/" estuary/post_install.sh
 		sudo cp estuary/post_install.sh $distro_dir/usr/bin/estuary/
 		sudo chmod 755 $distro_dir/usr/bin/estuary/post_install.sh
-		grep "/usr/bin/estuary/post_postinstall.sh" $distro_dir/etc/rc.local >/dev/null
+		if [ ! -f $distro_dir/etc/$rc_local_file ]
+			echo "$rc_local_template" | sudo tee $distro_dir/etc/$rc_local_file >/dev/null
+			sudo chown root:root $distro_dir/etc/$rc_local_file
+			sudo chmod 755 $distro_dir/etc/$rc_local_file
+		fi
+		
+		grep "/usr/bin/estuary/post_install.sh" $distro_dir/etc/$rc_local_file >/dev/null
 		if [ x"$?" != x"0" ]; then
-			sudo sed -i "/^exit/i/usr/bin/estuary/post_install.sh" $distro_dir/etc/rc.local
+			grep -E "^(exit)" $distro_dir/etc/$rc_local_file >/dev/null
+			if [ x"$?" = x"0" ]; then
+				sudo sed -i "/^exit/i/usr/bin/estuary/post_install.sh" $distro_dir/etc/$rc_local_file
+			else
+				sudo sed -i '$ a /usr/bin/estuary/post_install.sh' $distro_dir/etc/$rc_local_file
+			fi
 		fi
 
 		pushd $distro_dir/
